+++ src/PostgREST/Listener.hs
@@ -17,8 +17,11 @@ import           PostgREST.Version     (prettyVersion)
 import qualified PostgREST.AppState as AppState
 import qualified PostgREST.Config   as Config
 
-import Data.Either.Combinators (whenRight)
-import Protolude
+import           Control.Arrow             ((&&&))
+import           Data.Bitraversable        (bisequence)
+import           Data.Either.Combinators   (whenRight)
+import qualified Database.PostgreSQL.LibPQ as LibPQ
+import           Protolude
 
 -- | Starts the Listener in a thread
 runListener :: AppState -> IO ()
@@ -72,8 +75,8 @@ retryingListen appState = do
             -- reset the delay
             AppState.putNextListenerDelay appState 1
 
-          observer $ DBListenStart dbChannel
-
+          (pqHost, pqPort) <- SQL.withLibPQConnection db $ bisequence . (LibPQ.host &&& LibPQ.port)
+          observer $ DBListenStart pqHost pqPort dbChannel
           -- wait for notifications
           -- this will never return, in case of an error it will throw and be caught by onError
           forever $ SQL.waitForNotifications handleNotification db